+++ example/a.ts
@@ -1,6 +1,12 @@
-import { Elysia, t } from '../src'
-import { req } from '../test/utils'
+import { Elysia } from '../src'
 
-const app = new Elysia().get('/', () => 'Static Content')
+const app = new Elysia()
+	.ws('/ws/:id', {
+		message(ws, message) {
+			ws.send(message)
+		}
+	})
+	// .get('/ws/:id', () => 'hi')
+	.listen(3000)
 
-console.dir(app, { depth: 10 })
+// console.log(app.fetch.toString())