+++ src/utils.ts
@@ -46,10 +46,12 @@ export const replaceUrlPath = (url: string, pathname: string) => {
 
 export const isClass = (v: Object) =>
 	(typeof v === 'function' && /^\s*class\s+/.test(v.toString())) ||
-	// Handle import * as Sentry from '@sentry/bun'
-	// This also handle [object Date], [object Array]
-	// and FFI value like [object Prisma]
-	(v.toString().startsWith('[object ') &&
+	// Handle Object.create(null)
+	(v.toString &&
+		// Handle import * as Sentry from '@sentry/bun'
+		// This also handle [object Date], [object Array]
+		// and FFI value like [object Prisma]
+		v.toString().startsWith('[object ') &&
 		v.toString() !== '[object Object]') ||
 	// If object prototype is not pure, then probably a class-like object
 	isNotEmpty(Object.getPrototypeOf(v))